Struct jsonrpc_ws_server::ServerBuilder

pub struct ServerBuilder<M: Metadata, S: Middleware<M>> { /* fields omitted */ }

Builder for WebSockets server

Methods

impl<M: Metadata, S: Middleware<M>> ServerBuilder<M, S>

fn new<T>(handler: T) -> Self where T: Into<MetaIoHandler<M, S>>

Creates new ServerBuilder

fn event_loop_remote(self, remote: Remote) -> Self

Utilize existing event loop remote to poll RPC results.

fn session_meta_extractor<T: MetaExtractor<M>>(self, extractor: T) -> Self

Sets a meta extractor.

fn allowed_origins(self, allowed_origins: DomainsValidation<Origin>) -> Self

Allowed origins.

fn allowed_hosts(self, allowed_hosts: DomainsValidation<Host>) -> Self

Allowed hosts.

fn session_stats<T: SessionStats>(self, stats: T) -> Self

Session stats

fn request_middleware<T: RequestMiddleware>(self, middleware: T) -> Self

Sets a request middleware. Middleware will be invoked before each handshake request. You can either terminate the handshake in the middleware or run a default behaviour after.

fn start(self, addr: &SocketAddr) -> Result<Server, Error>

Starts a new WebSocket server in separate thread. Returns a Server handle which closes the server when droped.