Struct jsonrpc_ws_server::tokio_core::net::UdpFramed

pub struct UdpFramed<C> { /* fields omitted */ }

A unified Stream and Sink interface to an underlying UdpSocket, using the UdpCodec trait to encode and decode frames.

You can acquire a UdpFramed instance by using the UdpSocket::framed adapter.

Methods

impl<C> UdpFramed<C>

fn get_ref(&self) -> &UdpSocket

Returns a reference to the underlying I/O stream wrapped by Framed.

Note that care should be taken to not tamper with the underlying stream of data coming in as it may corrupt the stream of frames otherwise being worked with.

fn get_mut(&mut self) -> &mut UdpSocket

Returns a mutable reference to the underlying I/O stream wrapped by Framed.

Note that care should be taken to not tamper with the underlying stream of data coming in as it may corrupt the stream of frames otherwise being worked with.

fn into_inner(self) -> UdpSocket

Consumes the Framed, returning its underlying I/O stream.

Note that care should be taken to not tamper with the underlying stream of data coming in as it may corrupt the stream of frames otherwise being worked with.