Struct jsonrpc_ws_server::tokio_core::reactor::Interval

pub struct Interval { /* fields omitted */ }

A stream representing notifications at fixed interval

Intervals are created through the Interval::new or Interval::new_at methods indicating when a first notification should be triggered and when it will be repeated.

Note that timeouts are not intended for high resolution timers, but rather they will likely fire some granularity after the exact instant that they're otherwise indicated to fire at.

Methods

impl Interval

fn new(dur: Duration, handle: &Handle) -> Result<Interval, Error>

Creates a new interval which will fire at dur time into the future, and will repeat every dur interval after

This function will return a future that will resolve to the actual interval object. The interval object itself is then a stream which will be set to fire at the specified intervals

fn new_at(at: Instant,
dur: Duration,
handle: &Handle)
-> Result<Interval, Error>

Creates a new interval which will fire at the time specified by at, and then will repeat every dur interval after

This function will return a future that will resolve to the actual timeout object. The timeout object itself is then a future which will be set to fire at the specified point in the future.