Crate jsonrpc_ws_server

WebSockets server.

Re-exports

pub extern crate ws;

pub extern crate jsonrpc_core;

Modules

tokio_core
Future-powered I/O at the core of Tokio

Structs

CloseHandle
A handle that allows closing of a server even if it owned by a thread blocked in wait.

Error
The Error type.

Host
Host type

NoopExtractor
Dummy metadata extractor

Origin
Request Origin

RequestContext
Request context

Server
WebSockets server implementation.

ServerBuilder
Builder for WebSockets server

Enums

DomainsValidation
Specifies if domains should be validated.

ErrorKind
The kind of an error.

MiddlewareAction
Request middleware action

Traits

MetaExtractor
Metadata extractor from session data.

RequestMiddleware
Middleware to intercept server requests. You can either terminate the request (by returning a response) or just proceed with standard JSON-RPC handling.

SessionStats
Keeps track of open sessions

Type Definitions

Result
Convenient wrapper around std::Result.

SessionId
Session id