Crate jsonrpc_ws_server

WebSockets server.

Re-exports

pub use jsonrpc_core;

pub use ws;

pub use self::server_utils::tokio;

Structs

Broadcaster
A Broadcaster that can be used to send messages on all connections.

CloseHandle
A handle that allows closing of a server even if it owned by a thread blocked in wait.

Host
Host type

NoopExtractor
Dummy metadata extractor

Origin
Request Origin

RequestContext
Request context

Server
WebSockets server implementation.

ServerBuilder
Builder for WebSockets server

Enums

DomainsValidation
Specifies if domains should be validated.

Error
WebSockets Server Error

MiddlewareAction
Request middleware action

Traits

MetaExtractor
Metadata extractor from session data.

RequestMiddleware
Middleware to intercept server requests. You can either terminate the request (by returning a response) or just proceed with standard JSON-RPC handling.

SessionStats
Keeps track of open sessions

Type Definitions

Result
WebSockets Server Result

SessionId
Session id