Struct jsonrpc_ws_server::Server

pub struct Server { /* fields omitted */ }

WebSockets server implementation.

Methods

impl Server

pub fn addr(&self) -> &SocketAddr

Returns the address this server is listening on

pub fn start<M: Metadata, S: Middleware<M>>(
addr: &SocketAddr,
handler: Arc<MetaIoHandler<M, S>>,
meta_extractor: Arc<MetaExtractor<M>>,
allowed_origins: Option<Vec<Origin>>,
allowed_hosts: Option<Vec<Host>>,
request_middleware: Option<Arc<RequestMiddleware>>,
stats: Option<Arc<SessionStats>>,
remote: UninitializedRemote,
max_connections: usize
) -> Result<Server>

Starts a new WebSocket server in separate thread. Returns a Server handle which closes the server when droped.

impl Server

pub fn wait(self) -> Result<()>

Consumes the server and waits for completion

pub fn close(self)

Closes the server and waits for it to finish

pub fn close_handle(&self) -> CloseHandle

Returns a handle to the server that can be used to close it while another thread is blocking in wait.